% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glarma.R
\name{generate_sparse_cov_matrix}
\alias{generate_sparse_cov_matrix}
\title{Generate Sparse Covariance Matrix}
\usage{
generate_sparse_cov_matrix(
  n,
  p,
  sparse = TRUE,
  sparse_rho = 0.3,
  scale_power = 1
)
}
\arguments{
\item{n}{The dimension of the covariance matrix (number of rows and columns).}

\item{p}{The rank of the covariance matrix (number of non-zero eigenvalues).
Must be less than or equal to \code{n}.}

\item{sparse}{Logical indicating whether to enforce additional sparsity by
setting elements with absolute values below \code{sparse_rho} to zero.}

\item{sparse_rho}{Numeric threshold for enforcing sparsity.
Only used if \code{sparse} is \code{TRUE}.}

\item{scale_power}{The exponent used to scale the matrix elements,
with a default value of 1. This helps to control the magnitude
of the covariance matrix elements as \code{n} increases.}
}
\value{
A sparse \code{n} by \code{n} covariance matrix with rank \code{p}.
If \code{sparse} is \code{TRUE}, elements with absolute values below
\code{sparse_rho} are set to zero to increase sparsity,
while ensuring that the matrix is at least semi-definite.
}
\description{
Generates a sparse covariance matrix with specified dimension and rank.
The generated matrix is scaled to control the magnitude of its elements
and can be further sparsified based on a given threshold.
}
\examples{
# Generate a 10x10 sparse covariance matrix with rank 5
sparse_cov_matrix <- generate_sparse_cov_matrix(n = 10, p = 5)

# Generate a sparser matrix with elements below 0.3 set to zero
sparser_cov_matrix <- generate_sparse_cov_matrix(n = 100, p = 50,
                                                 sparse = TRUE, sparse_rho = 0.3)

# Generate a matrix with a different scale power
sparse_cov_matrix <- generate_sparse_cov_matrix(n = 100, p = 50, scale_power = 2)

}
